refactor(compaction): extract compaction into its own package
Move CountCompactions from internal/usage to internal/compaction.
This decouples transcript parsing from the usage API, preparing
for the stdin-first architecture in v1.
